Move tests from pkg/serializer/internal/metrics to pkg/metrics
These tests check the behavior of metrics.IterableSeries and not serializer.internal.metrics.IterableSeries
